Remove minimum stack size in FreeRTOS Thread
A previous change made it so that a Thread object on FreeRTOS always had
a stack that was at least 4096 bytes even if fewer bytes were requested.
No justification was provided alongside that change and in testing, eRPC
seems to operate correctly without enforcing this minimum stack size, so
this change is being reverted.